server/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.16
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
fsspec==2024.9.0
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
lightfm==1.17
MarkupSafe==2.1.5
mpmath==1.3.0
numpy==1.24.4
packaging==24.1
pandas==2.0.3
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sqlparse==0.5.1
sympy==1.13.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
